Fix #39

+++ App/Model/Directory.php
@@ -82,9 +82,7 @@ class Directory {
 					continue;
 				} elseif ( $fileinfo->isDir() ) {
 					if ( static::_remove_children( $path ) && static::_is_removable( $path ) ) {
-						if ( ! rmdir( $path ) ) {
-							throw new \Exception( sprintf( '[Snow Monkey Forms] Can\'t remove directory: %1$s.', $path ) );
-						}
+						static::remove( $path );
 					}
 				} elseif ( $fileinfo->isFile() ) {
 					if ( static::_is_removable( $path ) ) {
@@ -103,19 +101,14 @@ class Directory {
 	public static function remove( $file ) {
 		$fileinfo = new SplFileInfo( $file );
 
-		try {
-			if ( $fileinfo->isFile() ) {
-				if ( ! unlink( $file ) ) {
-					throw new \Exception( sprintf( '[Snow Monkey Forms] Can\'t remove file: %1$s.', $file ) );
-				}
-			} elseif ( $fileinfo->isDir() ) {
-				if ( ! rmdir( $file ) ) {
-					throw new \Exception( sprintf( '[Snow Monkey Forms] Can\'t remove directory: %1$s.', $file ) );
-				}
+		if ( $fileinfo->isFile() && is_writable( $file ) ) {
+			if ( ! unlink( $file ) ) {
+				throw new \RuntimeException( sprintf( '[Snow Monkey Forms] Can\'t remove file: %1$s.', $file ) );
 			}
-		} catch ( \Exception $e ) {
-			error_log( $e->getMessage() );
-			return false;
+		} elseif ( $fileinfo->isDir() && is_writable( $file ) ) {
+			if ( ! rmdir( $file ) ) {
+				throw new \RuntimeException( sprintf( '[Snow Monkey Forms] Can\'t remove directory: %1$s.', $file ) );
+			}
 		}
 
 		return true;
@@ -137,22 +130,17 @@ class Directory {
 			return true;
 		}
 
-		try {
-			$handle = fopen( $htaccess, 'w' );
-			if ( ! $handle ) {
-				throw new \Exception( '[Snow Monkey Forms] .htaccess can\'t create.' );
-			}
+		$handle = fopen( $htaccess, 'w' );
+		if ( ! $handle ) {
+			throw new \RuntimeException( '[Snow Monkey Forms] .htaccess can\'t create.' );
+		}
 
-			if ( false === fwrite( $handle, "Deny from all\n" ) ) {
-				throw new \Exception( '[Snow Monkey Forms] .htaccess can\'t write.' );
-			}
+		if ( false === fwrite( $handle, "Deny from all\n" ) ) {
+			throw new \RuntimeException( '[Snow Monkey Forms] .htaccess can\'t write.' );
+		}
 
-			if ( ! fclose( $handle ) ) {
-				throw new \Exception( '[Snow Monkey Forms] .htaccess can\'t close.' );
-			}
-		} catch ( \Exception $e ) {
-			error_log( $e->getMessage() );
-			return false;
+		if ( ! fclose( $handle ) ) {
+			throw new \RuntimeException( '[Snow Monkey Forms] .htaccess can\'t close.' );
 		}
 
 		return true;
